+++ lib/new_relic/agent/configuration/default_source.rb
@@ -1559,7 +1559,7 @@ module NewRelic
           :type => String,
           :dynamic_name => true,
           :allowed_from_server => false,
-          :description => 'Controls auto-instrumentation of the aws_sdk_firehose library at start-up. May be one of `auto`, `prepend`, `chain`, `disabled`.'
+          :description => 'Controls auto-instrumentation of the aws-sdk-firehose library at start-up. May be one of `auto`, `prepend`, `chain`, `disabled`.'
         },
         :'instrumentation.aws_sdk_lambda' => {
           :default => 'auto',